django/requirements.txt
Django>=3.2,<4
gunicorn
uvicorn
djangorestframework
django-filter
dj-database-url
